% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\docType{methods}
\name{control}
\alias{control}
\alias{control,nblda-method}
\alias{control,nblda_trained-method}
\title{Accessors for the 'control' slot.}
\usage{
\S4method{control}{nblda}(object)

\S4method{control}{nblda_trained}(object)
}
\arguments{
\item{object}{an \code{nblda} or \code{nblda_trained} object.}
}
\description{
This slot stores control parameters for training NBLDA model.
}
\examples{
\dontrun{
1L
}

}
\seealso{
\code{\link{trainNBLDA}}
}
