% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{nblda-class}
\alias{nblda-class}
\title{\code{nblda} object}
\description{
This object is the main class for NBLDA package. It stores inputs, results and call info for the trained model.
}
\details{
Objects can be created by calls of the form \code{new("nblda", ...)}. This type
of objects is created as a result of \code{trainNBLDA} function of \code{NBLDA} package.
It is then used in \code{predict} function for predicting class labels of new samples.
}
\section{Slots}{


\describe{
  \item{\code{input}:}{an \code{nblda_input} object including the count matrix (or data.frame) and class labels.}
  \item{\code{result}:}{an \code{nblda_trained} object with elements from corss-validated and final models.}
  \item{\code{call}:}{a call expression.}
}
}

\seealso{
\code{\linkS4class{nblda_trained}}, \code{\linkS4class{nblda_input}}
}
\author{
Dincer Goksuluk
}
