% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{nblda_input-class}
\alias{nblda_input-class}
\title{\code{nblda_input} object}
\description{
This object is the subclass for NBLDA package. It stores input objects, i.e. count data and class labels.
}
\note{
nblda_input icerisinde yer alan slotlarin aciklamalari burada verilecek.
}
\section{Slots}{


\describe{
  \item{\code{x}:}{a data.frame or matrix. Count data input for NBLDA classifier.}
  \item{\code{y}:}{a vector of length equal to number of rows of x. This is the class labels of each subject.
  Should be either a numeric vector or factor.}
}
}

\author{
Dincer Goksuluk
}
