% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\docType{methods}
\name{nbldaTrained}
\alias{nbldaTrained}
\alias{nbldaTrained,nblda-method}
\alias{nbldaTrained,nblda_trained-method}
\title{Accessors for the 'crossValidated' slot.}
\usage{
\S4method{nbldaTrained}{nblda}(object)

\S4method{nbldaTrained}{nblda_trained}(object)
}
\arguments{
\item{object}{an \code{nblda} or \code{nblda_trained} object.}
}
\description{
This slot stores the results for cross-validated model, e.g tuning results, optimum model parameters etc.
}
\examples{
\dontrun{
1L
}

}
\seealso{
\code{\link{trainNBLDA}}
}
