% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\docType{methods}
\name{normalization}
\alias{normalization}
\alias{normalization,nblda-method}
\alias{normalization,nblda_trained-method}
\title{Accessors for the 'type' slot.}
\usage{
\S4method{normalization}{nblda}(object)

\S4method{normalization}{nblda_trained}(object)
}
\arguments{
\item{object}{an \code{nblda} or \code{nblda_trained} object.}
}
\description{
This slot stores the name of normalization method. Normalization is defined using \code{type} argument in \code{\link{trainNBLDA}} function.
}
\examples{
\dontrun{
1L
}

}
\seealso{
\code{\link{trainNBLDA}}
}
