% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\docType{methods}
\name{print}
\alias{print}
\alias{print.nblda}
\alias{print,nblda-method}
\title{Print Method for the \code{nblda} Class}
\usage{
\method{print}{nblda}(x, ...)

\S4method{print}{nblda}(x, ...)
}
\arguments{
\item{x}{an object of class \code{nblda} to be printed.}

\item{\dots}{further arguments to be passed to \code{\link[base]{print}}.}
}
\description{
Print the results of an object returned from \code{\link{trainNBLDA}}.
}
\examples{
\dontrun{
set.seed(2128)
counts <- generateCountData(n = 20, p = 10, K = 2, param = 1, sdsignal = 0.5, DE = 0.8,
                            allZero.rm = FALSE, tag.samples = TRUE)
x <- t(counts$x + 1)
y <- counts$y
xte <- t(counts$xte + 1)
ctrl <- control(folds = 2, repeats = 2)

fit <- trainNBLDA(x = x, y = y, type = "mle", tuneLength = 10,
                  metric = "accuracy", train.control = ctrl)

print(fit)
}

}
\author{
Dincer Goksuluk
}
