% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\docType{methods}
\name{show}
\alias{show}
\alias{show.nblda}
\alias{show,nblda-method}
\title{Show Method for the \code{nblda} Class}
\usage{
\method{show}{nblda}(object)

\S4method{show}{nblda}(object)
}
\arguments{
\item{object}{an object of class \code{nblda} to be printed.}
}
\description{
Show the results of an object returned from \code{\link{trainNBLDA}} on R console.
}
\examples{
\dontrun{
set.seed(2128)
counts <- generateCountData(n = 20, p = 10, K = 2, param = 1, sdsignal = 0.5, DE = 0.8,
                            allZero.rm = FALSE, tag.samples = TRUE)
x <- t(counts$x + 1)
y <- counts$y
xte <- t(counts$xte + 1)
ctrl <- control(folds = 2, repeats = 2)

fit <- trainNBLDA(x = x, y = y, type = "mle", tuneLength = 10,
                  metric = "accuracy", train.control = ctrl)

show(fit)
}

}
\author{
Dincer Goksuluk
}
