% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{control-NBLDA}
\alias{control-NBLDA}
\alias{control}
\title{Control parameters for trained NBLDA model.}
\usage{
control(folds = 5, repeats = 2, foldIdx = NULL, rhos = NULL, beta = 1,
  prior = NULL, transform = FALSE, alpha = NULL, truephi = NULL,
  target = 0, phi.epsilon = 0.15, normalize.target = FALSE,
  delta = NULL, return.selected.gene.names = FALSE, multicore = FALSE,
  ...)
}
\arguments{
\item{folds}{A positive integer. The number of folds for k-fold model validation.}

\item{repeats}{A positive integer. This is the number of repeats for k-fold model validation. If NULL, 0 or negative, it is set to 1.}

\item{foldIdx}{a list with indices of hold-out samples for each fold. It should be a list where folds are nested within repeats.
If NULL, \code{folds} and \code{repeats} are used to define hold-out samples.}

\item{rhos}{A vector of tuning parameters that control the amount of soft thresholding performed. If NULL, it is automatically generated within
\code{\link{trainNBLDA}} using \code{tuneLength}, i.e. the length of grid search. See details.}

\item{beta}{A smoothing term. A Gamma(beta,beta) prior is used to fit the Poisson model. Recommendation is to just leave it at 1, the default value.
See Witten (2011) and Dong et. al. (2016) for details.}

\item{prior}{A vector of length equal to the number of classes indicating the prior class probabilities. If NULL, all classes are assumed to be
equally distributed.}

\item{transform}{a logical. If TRUE, count data is transformed using power transformation. If \code{alpha} is not specified the power transformation
parameter is automatically calculated using goodness-of-fit test. See Witten (2011) for details.}

\item{truephi}{a vector of length equal to the number of variables representing the true overdispersion parameters for each variable. If a single
value is given, it is replicated for all variables. If a vector of length unequal to the number of variables is given, the first element of this
vector is used and replicated for all variables. If NULL, estimated overdispersions are used in the classifier. See details.}

\item{phi.epsilon}{a positive value for controlling the number of features whose dispersions are shrinked towards 0. See details.}

\item{apha}{a numeric value within [0, 1] to be used for power transformation.}
}
\value{
give information about returned objects here.
}
\description{
Define control parameters to be used within \code{\link{trainNBLDA}} function.
}
\details{
\code{rhos} is used to control the level of sparsity, i.e. the number of variables (or features) used in classifier. If a variable has
no contribution to discrimination function, it should be removed from the model. By setting rhos within the interval [0, Inf], it is possible
control the amount of variables that is removed from the model. As the upper bound of rhos decreases towards 0, fewer variables are removed.
If rhos = 0, all variables are included in classifier.

\code{truephi} controls how Poisson model differs from Negative Binomial model. If overdispersion is zero, Megative Binomial model converges to
Poisson model. Hence, the results from \code{\link{trainNBLDA}} is identical to PLDA results from \code{\link[PoiClaClu]{Classify}} when truephi = 0.

\code{phi.epsilon} is a value used to shring estimated overdispersions towards 0. Poisson model assumes that there is no overdispersion in the
observed counts. However, this is not a valid assumption in highly overdispersed count data. \code{NBLDA} performs a shrinkage on estimated
overdispersions. Although the amount of shrinkage is dependent on several parameters such as \code{delta}, \code{target} and \code{truephi}, some
of the shrinked overdispersions might be very close to 0. By defining a threshold value for shrinked overdispersions, it is possible to shrink
very small overdispersions towards 0. If estiamted overdispersion is below \code{phi.epsilon}, it is shrinked to 0.
}
\examples{
# Enter examples here.
1L

}
\references{
Witten, DM (2011). Classification and clustering of sequencing data using a Poisson model.
Ann. Appl. Stat. 5(4), 2493--2518. doi:10.1214/11-AOAS493.

Dong, K., Zhao, H., Tong, T., & Wan, X. (2016). NBLDA: negative binomial linear discriminant analysis for RNA-Seq data.
BMC Bioinformatics, 17(1), 369. http://doi.org/10.1186/s12859-016-1208-1.
}
\seealso{
<add seealso info here>
}
\author{
Turcosa Developing Team
}
\keyword{keywords_here}
