% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{nblda_trained-class}
\alias{nblda_trained-class}
\title{\code{nblda_trained} object}
\description{
This object is the subclass for NBLDA package. It stores cross-validated results and the final model.
}
\section{Slots}{


\describe{
  \item{\code{crossValidated}:}{a list. Returns the results from cross-validation.}
  \item{\code{finalModel}:}{a list with elements from final model using optimum model parameters from cross-validated model.}
  \item{\code{control}:}{a list with controlling parameters for fitting NBLDA classifier.}
}
}

\author{
Dincer Goksuluk
}
